Completed Sign Up Page and Fixes in Login Page
Fixes in Login Page: 1) Added Focus Nodes on every text field,  2) init and dispose functions for text controllers and focus nodes as shown in documentation, 3) Added clear icon on e-mail text box, 4) If the keyboard is enabled and the space outside text boxes is tapped, it will be hidden, 5)Icons used instead of images, so some of them will have to be removed later. All the above have been implemented in the signup page as well.